backend/requirements.txt
FLASK==2.1.2
psycopg2-binary==2.9.1
python-dotenv==1.0.0
flask_cors==3.0.10
pytz==2023.3